requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
sarvamai
jinja2==3.1.4
python-multipart==0.0.12
twilio>=9.0.0
pyngrok>=8.0.0
